drop the 2 from date_parse2
the original, date_parse from alpine-mbox, which was simpler but much
slower due to its use of strptime, scanf, and timegm, is not used here
in any case.

there might be a historical reason to keep the 2, you know, as a legacy
to the kids about the ancestry or something.  probably i would actually
do that too, if i weren't touching the function, but as it is i'm
de-C++-izing it and touching it up to work specifically for http date
headers, rather than mbox timestamps.  so it almost ought to be
date_parse3, but then it just starts to get confusing.  putting the
original in git history here at least leaves the story for the really
curious.