parse last-modified timestamp
I'm half considering calling last_modified a long here, instead of
a time_t.  Just checked - the tv_sec field in struct timespec (which
is used in the struct stat that we need to fill for the fuse getattr
call) is a time_t, so i guess we can leave it a time_t.