export pfxmatch
... Probably deserves to be in a separate util file, but we'll leave
it for now.  The idea with leaving it in cheddar as a static method
(besides laziness) was mainly to allow inlining and, i was supposing,
to allow strlen() be evaluated as a compile-time constant for the
header name string literals.  And maybe we still get that by leaving
it in cheddar, despite also exporting it.