export addpath
might want to rename it something more specific to hyperfs-cache...

doing this for a way to stash other strings - in this case, the use
case is link targets.  I expect this is faster than malloc, and doesn't
require tracking all the allocations, which i expect to be basically
permanent / last the lifetime of the process.