introduce xasprintf
i'd like the idea of making this a proper function, but i do not feel
like doing stdarg processing, so, a macro it is.

note: i am generally thinking to use our own hyperfs-cache addpath
mechanism for allocating strings (especially for strings which are
expected to be valid until the end of the program) ... but i think
this may be handy for a few items at startup.

NOTE: _GNU_SOURCE should be defined at the top of any source file
that uses xasprintf, before any headers are included (even this one).
The xasprintf.h doesn't itself define _GNU_SOURCE, because i don't
want to introduce any awkward rules about our headers having to be
included in a particular order.