hyperfs_readlink
from watching the logs i kinda "trust the system" enough not to
bother about the details.  If hyperfs_readlink is called, assume
getattr has already been called and this is a link and the link
target thus has already been stored in the cache.

haven't actually tested this yet - in theory symlinks mean either
(1) the Location for a a Moved Permanently/Found, or (2) an href
that is not a relative, single path component.