be more careful in expand_ministat
mst->mtime is a union member (doubles as a pointer to symlink target)
so we don't want to set the times to it in this case.  mst->size may
end up similarly.  also we can leave fields unset to default to zero,
because we do an initial memset.  so let's leave size at 0 for dirs
(it's proper to set it for regular files and symlinks), and for now
let's only set times for regular files.  maybe later we'll set it for
dirs too (from the parent dir's listing), but definitely we should not
be setting times for symlinks.