ministat: add hyperdir field for directory listing
this is a char pointer to a flat list of directory entries, terminated
by a null string.  entries that refer to directories "sub/" will be
stored with the slash *at the beginning* rather than at the end.  There
are two reasons for this: 1. to determine directory or not immediately
by examining the first character, rather than scanning the string with
strlen to find the last character; and 2. so that &p[1] (for subdir
entries) is a ready-to-go nul-terminated string to pass along to
readdir.

Initially I had in mind to make a struct hyperdir with a count of
entries, and possibly allow each entry to contain type info ... but i am
leaning to think that the optional '/' prefix is sufficient.

(Well, except for symlinks.  I wonder if they deserve special treatment.
But i haven't encountered any yet and i don't especially care about them.
We'll see.)