add escape methods
eventually we'll want these in order to handle filenames with
unpleasant characters in them.

the switch cases in needs_escape are based on experimentation
to see what apache does with variously named files.

curiously, apache presents a '&' as '&amp;', but everything else it
encodes with %XX (%%%02x).  So we need to know how to decode that, too;
but it's still OK for us to encoded '&' with %XX (that works for any
character).

I do have in mind to make a 256 byte lookup table with needs_escape,
though i doubt the performance gain will be measurable.