add add_pathlen
like add_pathbuf, but tack on a NUL terminator, too

wanted to have this call add_pathbuf, and then just add the NUL
afterwards, but the 'does it fit?' check should be done just once
with a single call to get_pathbuf.  so we end up with two very
similar looking functions.  oh well.