introduce hyperget
feels like getrange or get_head_info, but this time ask for the whole
file (GET) and leave the stream ready to start reading it.

feels a bit bloated, if i'm honest.  possibly with all the error
handlng, idk.  might be nice to break it up or something.

anyway, this will be used for reading the automatic index pages
for /dir/ paths.