sendo.h: introduce SENDO, S_ENDO
make it easier to stuff sent to socket

SENDO is per-line now, not the whole command/header block.  I think
this is OK, since it's not actually written to the socket until
fflush.

S_ENDO is a bit of a sad hack due to the limitation of __VA_ARGS__
macro expansion.  but whatever we'll live.