add -s option for single-threaded fs
i was seeing multiple socket connections getting opened at the same
time, which i assume is not actually what we want, especially once
we try to re-use the same connection with keep-alive.

adding an extra arg means we can't just shuffle stuff around in main's
argv anymore; we actually need more space.  60 args ought to be plenty
(4 should be enough, really), and i guess i'm not feeling the need
to malloc a more exact sized array.

also, move the now-unused shift_n_push to _spare-parts.  yes it was fun,
but we don't need it anymore.