hyperfs-cache: add ptrbuf and friends
not clear if we'll actually need this, but it seemed like an interesting
idea - a separate allocator for pointers.

ptrbuf etc could (maybe should) be generic void **'s, but that requires
casting when actually using them (which is annoying), and likely these
will be used for char pointers anyway.

originally i had in mind to use this for, eg, the fuse_args array, but i
ended up going with a stack buffer for that.