add initial tstamp parsing to readdir
not doing anything useful with it yet, but at least we dropped a note
in there.

also add tz_offset and tz_offset_found to hyperfs-state.  the idea here
is to set these appropriately once a regular file is found with a proper
last-modified http header, that can be matched to the timestamp from the
index page.  then this server tz offset can be applied to all future dir
timestamps, which come from their parent dir's index page but do not
have their own last-modified timestamp.